+++ libft/ft_memmove.c
@@ -2,9 +2,9 @@
 
 void	*ft_memmove(void *dst, const void *src, size_t len)
 {
-	char	*s;
-	char	*d;
-	int		i;
+	char *s;
+	char *d;
+	size_t i;
 
 	s = (char *)src;
 	d = (char *)dst;
@@ -13,7 +13,7 @@ void	*ft_memmove(void *dst, const void *src, size_t len)
 	if (d > s)
 	{
 		i = 1;
-		while (i < (int)len + 1)
+		while (i < len + 1)
 		{
 			d[len - i] = s[len - i];
 			i++;
@@ -22,7 +22,7 @@ void	*ft_memmove(void *dst, const void *src, size_t len)
 	else
 	{
 		i = 0;
-		while (i < (int)len)
+		while (i < len)
 		{
 			d[i] = s[i];
 			i++;
@@ -30,22 +30,3 @@ void	*ft_memmove(void *dst, const void *src, size_t len)
 	}
 	return (d);
 }
-
-// #include<string.h>
-
-// int main() {
-//     char src[] = "Hello, World!";
-//     char dest[20];
-
-//     // テストケース 1: オーバーラップなしのコピー
-//     // ft_memmove(dest, src, strlen(src) + 1);
-//     memmove(dest, src, strlen(src) + 1);
-//     printf("Test Case 1: %s\n", dest);  // 期待される出力: "Hello, World!"
-
-//     // テストケース 2: src と dest がオーバーラップ
-//     memmove(src + 3, src, strlen(src) + 1);
-//     // ft_memmove(src + 3, src, strlen(src) + 1);
-//     printf("Test Case 2: %s\n", src + 3);  // 期待される出力: "lo, World!"
-
-//     return (0);
-// }
